+++ CHANGELOG.md
@@ -1,3 +1,6 @@
-## [0.0.1] - TODO: Add release date.
+## [0.0.1] - 2020-07-13.
 
-* TODO: Describe initial release.
+* Added timewheel widget to allow input of time value.
+* returns selected time value in 24 hour format but displays in 12 hour format
+* initial selected value equal to DateTime.now() value
+* customizable height, width, fontsize, itemExtent and spacing of dividers